+++ osu.Server.PerformanceCalculator/Commands/ScoresCommand.cs
@@ -10,8 +10,8 @@ namespace osu.Server.PerformanceCalculator.Commands
     public class ScoresCommand : CalculatorCommand
     {
         [Argument(0, "scores", "The score IDs to recompute performance for.")]
-        public int[] Scores { get; set; } = null!;
+        public long[] Scores { get; set; } = null!;
 
-        protected override IEnumerable<int> GetScores() => Scores;
+        protected override IEnumerable<long> GetScores() => Scores;
     }
 }
